Add animation duration to properties
This is useful when an `AnimationPlayer` controls the Spine node
and you need to tell your animation how long it's going to be. With
the wrong duration, the animation will be broken and there's no
better way to check this than having it visible.

It would be nicer if it was truly read-only. Now it looks like it
can be edited, though it can't and obviously shouldn't.